CSE546-FallA2021-master/requirements.txt
boto3
numpy
pillow
requests
numpy
